:sparkles: Make Book Filament Resource

+++ database/migrations/2024_08_30_125623_create_books_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('slug');
             $table->string('isbn');
             $table->string('genre');
-            $table->float('price');
+            $table->integer('price'); // in cents
             $table->text('description');
             $table->timestamps();
         });
